Update README.md
Project on Hackaday:
https://hackaday.io/project/170728-elevated-body-temperature-detector-cloud-enabled

The system constantly runs and checks for movement in the area. If movement is detected, sonic sensors estimate the distance to a closest target and IR sensor takes IR readings. If IR readings from all four sensors are higher than a threshold "EBT" is reported. The system also runs and post data to the web-page when no motion is detected or when only motion is detected (no EBT)